+++ Dockerfile
@@ -2,7 +2,7 @@ FROM python:3.5-alpine
 
 RUN apk add --no-cache nmap nmap-scripts git
 
-RUN pip install --no-cache-dir xmltodict google-cloud-storage boto3
+RUN pip install --no-cache-dir xmltodict google-cloud-storage boto3 neo4j
 
 RUN git clone https://github.com/vulnersCom/nmap-vulners /usr/share/nmap/scripts/vulners && nmap --script-updatedb
 RUN mkdir /shared
